Legend: Fix SetGeoView not subscribing to SceneView scene changes
`LegendDataSource.SetGeoView(...)` tried to find SceneProperty on MapView type instead of on SceneView type.  This caused `DependencyPropertyDescriptor.FromProperty` to create a new "attached" PropertyDescriptor, instead of looking up the correct one.  As a result, GeoViewDocumentChanged listener was not actually called.  This PR fixes the type and event fires as expected.