Fixes issue with bookmark loading (#297)
Sometimes, when the Map is set on a MapView, the map property change event happens before it can be subscribed to. To work around this, Xamarin platforms were attempting to load the map in all cases when the geoview is set. That step is also required on UWP (and probably WPF), but it was missed because of a deficiency in the bookmarks tests for UWP & WPF.

This commit fixes that issue and updates the test so that the original issue would have been found.