Updates to build configuration for usability
Whenever an Android project is selected for build, Visual Studio resets the configuration to AnyCPU. Forms UWP was set to build but not deploy UWP when AnyCPU was selected. In effect, Forms UWP would build, but then the old/last-deployed version would be started, which led to confusing/unexpected behavior.