Forms - Update Legend to use CollectionView
This change improves the appearance of Legend on Forms UWP. Previously, there were hover and click effects when moving the mouse over the legend. This behavior was misleading, because the legend is not interactive.

This change replaces the ListView with a CollectionView, and updates the default templates to return views instead of cells.

Note that while this change won't break the build, there is an exception at run if users have customized the data template properties with templates that consist of cells, rather than views.