Initialization from GeoView improvements
This change ensures that `InitializeTimePropertiesAsync` taking a GeoView loads the GeoView's map or scene before checking layer time properties. This ensures that layers that can be loaded are loaded.

Also, an error was fixed where if a geoview's time extent was already null, the slider's current extent would not be set.