TimeSlider improvements (#405)
* Update WPF Timeslider sample

Sample now demonstrates broader range of functionality, and also properly uses TimeSlider events to meet user expectations.

Note that the setting to use da-DK as the language was removed; this was causing unexpected date formatting behavior on en-US (and other) systems.

* Add UWP TimeSlider sample

* Add Forms sample and other sample improvements

* Native iOS TimeSlider sample

* Add Android TimeSlider sample

* TimeSlider (all): Guard against invalid time steps

Previously, there would be strange behavior with time step values less than one.

* TimeSlider markdown doc update

* TimeSlider WPF: remove background behind Play

* WPF: fix layout issue affecting track, thumb

* Mitigate effects of imprecision for intervals

For some time step counts, there would previously be an issue where the last interval would be dropped, despite being a close match. Rounding from certain values was causing extra milliseconds to break the time comparison. This change adjusts the time division code to use more precise time intervals by default, round the resulting value, and snap the last time step to the end time if the difference is less than 10 minutes.

So far this gives good results with the time slider as implemented, but I'm concerned this is just papering over other defects.

* Fix TimeSlider Forms API Ref

* Misc. fixes

* Initialization from GeoView improvements

This change ensures that `InitializeTimePropertiesAsync` taking a GeoView loads the GeoView's map or scene before checking layer time properties. This ensures that layers that can be loaded are loaded.

Also, an error was fixed where if a geoview's time extent was already null, the slider's current extent would not be set.

* Changes from review

* Revise time snapping behavior

Changes:

- Reverts change to how time division works
- When adding time steps, now adds a final time step to ensure that the full time extent is accessible. In cases where the size of the gap is less than half a time step, the last step is replaced with the end time.

* Don't always default to instantaneous time extent

This was leading to bad behavior in the sample data. New behavior better matches JS default.

* Fix HTTP url in samples