Update ARCore, restore Android hit test behavior
Due to a bug in the ARCore 1.26.0 Xamarin binding, hit test filtering had to be removed, which led to less accurate surface placement when there were multiple planes in view. This issue has been fixed in the 1.29.0 release.

This PR/commit updates to 1.29.0 and restores the previous behavior.

I've tested both the Forms and native Android AR samples to confirm the behavior has been improved and everything else is working as expected.