Updates FloorFilter with Indoors support
This PR updates the core floorfilter controller with support for applying extended location information. It also updates the MAUI and windows UI implementations to take advantage of this feature and adds a sample feature for testing.

Implementation will only set the floor if the incoming point is within the currently selected facility's bounding rectangle.

Sample supports testing with empty and invalid floor information.