BugFix: BasemapGallery fails to load default Basemaps in WPF (#586)
* Optimize BasemapGallery initialization

Removed redundant initialization of `AvailableBasemaps` in the `BasemapGallery` constructor. Added a null check in `BasemapGallery_Loaded` to initialize `AvailableBasemaps` only if it is null, calling `_controller.LoadFromDefaultPortal()` asynchronously if needed. This change prevents unnecessary initialization and allows for external setting of `AvailableBasemaps` before the control is loaded.

* Initialize AvailableBasemaps and improve load check

Add initialization flag for BasemapGallery collection